feat(cost-tracker): implement sign-out functionality with dedicated button component
